requirements.txt
alembic==1.6.5
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
DateTime==4.3
fastapi==0.65.2
greenlet==1.1.0
h11==0.12.0
idna==2.10
importlib-metadata==4.5.0
Mako==1.1.4
MarkupSafe==2.0.1
minio==7.0.3
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.16.0
slugify==0.0.1
SQLAlchemy==1.4.17
SQLAlchemy-Utils==0.37.7
timestamp==0.0.1
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
zipp==3.4.1
zope.interface==5.4.0
pytest~=6.2.4
mock==4.0.3
pytest-cov==2.8.1
typer==0.3.2
pyhumps==3.0.2
pandas==1.2.5
prometheus-client==0.7.1
starlette==0.14.2
aiofiles==0.7.0
XlsxWriter==1.4.4
openpyxl==3.0.7
numpy~=1.21.0
openpyxl==3.0.7
cachetools~=4.2.2
fakeredis~=1.6.0
pytest-asyncio==0.15.1
enum34~=1.1.10
aiohttp~=3.7.4.post0
asyncio==3.4.3
grpcio==1.39.0
git+https://nghia.nd:zb4BrpA6F7V6MeREy3jx@git.teko.vn/platform/tekone.git
pyarrow==5.0.0
redis==3.5.3
aioredis==1.3.1
teko-export-test
opentelemetry-exporter-jaeger==1.7.1
opentelemetry-instrumentation-fastapi==0.26b1
opentelemetry-instrumentation-requests==0.26b1
opentelemetry-instrumentation-sqlalchemy==0.26b1
opentelemetry-sdk==1.7.1
opentelemetry-api==1.7.1
xlrd==2.0.1
starlette==0.14.2
starlette-context==0.3.3
freezegun==1.2.0
kafka-python==2.0.2
unidecode==1.3.4
deepdiff==5.8.1